добавил модуль utils.js
